+++ hangongsu-api/src/main/java/com/hangongsu/api/HangongsuApplication.java
@@ -1,0 +1,16 @@
+package com.hangongsu.api;
+
+import com.hangongsu.support.common.config.EnablePropertiesConfig;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
+
+@Import(EnablePropertiesConfig.class)
+@SpringBootApplication
+public class HangongsuApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HangongsuApplication.class, args);
+    }
+
+}
