+++ hangongsu-api/src/main/java/com/hangongsu/api/advice/GlobalExceptionHandler.java
@@ -1,0 +1,24 @@
+package com.hangongsu.api.advice;
+
+import com.hangongsu.api.response.ErrorResponse;
+import com.hangongsu.support.exception.ErrorCode;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+    private static final String LOG_FORMAT = "Class :: {} | Code :: {} | Message :: {}";
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleException(Exception e) {
+        ErrorCode errorCode = ErrorCode.INTERNAL_SERVER_ERROR;
+        ErrorResponse response = ErrorResponse.of(errorCode);
+
+        log.error(LOG_FORMAT, e.getClass().getSimpleName(), errorCode.getCode(), errorCode.getMessage());
+
+        return ResponseEntity.status(errorCode.getStatus()).body(response);
+    }
+}
