+++ hangongsu-api/src/main/java/com/hangongsu/api/provider/auth/OAuthRedirectUriProvider.java
@@ -1,0 +1,9 @@
+package com.hangongsu.api.provider.auth;
+
+import com.hangongsu.core.domain.user.vo.Platform;
+
+import java.net.URI;
+
+public interface OAuthRedirectUriProvider {
+    URI generateRedirectUri(Platform platform);
+}
