+++ hangongsu-core/src/main/java/com/hangongsu/core/application/auth/OAuthLoginService.java
@@ -1,0 +1,5 @@
+package com.hangongsu.core.application.auth;
+
+public interface OAuthLoginService {
+
+}
