+++ hangongsu-infrastructure/src/main/java/com/hangongsu/infratstructure/client/auth/OAuthClient.java
@@ -1,0 +1,8 @@
+package com.hangongsu.infratstructure.client.auth;
+
+public interface OAuthClient {
+
+    String getAccessToken(String authorizationCode);
+
+    String retrieveUserInfo(String accessToken);
+}
