+++ hangongsu-support/build.gradle
@@ -1,0 +1,11 @@
+dependencies {
+}
+
+jar {
+    enabled = true
+}
+
+bootJar {
+    enabled = false
+}
+
