--- src/main/java/com/hangongsu/matching/HangongsuApplication.java
@@ -1,13 +1,0 @@
-package com.hangongsu.matching;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HangongsuApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HangongsuApplication.class, args);
-    }
-
-}
