--- src/test/java/com/hangongsu/matching/HangongsuApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hangongsu.matching;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HangongsuApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
